restructuring of the codebase
