+++ python/mdvtools/auth/authutils.py
@@ -1,10 +1,226 @@
-from flask import jsonify, session, request, current_app
+from flask import jsonify, session, request,redirect, current_app
 from functools import wraps
+
+# in_memory_cache.py
+user_cache = {}  # key: auth0_id -> user details
+user_project_cache = {}  # key: user_id -> project permissions
+all_users_cache = []  # list of all user summaries
+active_projects_cache = []
+
+def is_authenticated():
+    """Validate the current user via Auth0 or Shibboleth."""
+    ENABLE_AUTH = current_app.config.get("ENABLE_AUTH", False)
+    
+    if not ENABLE_AUTH:
+        return True
+
+    auth_method = session.get("auth_method")
+    if auth_method == "auth0":
+        user, error_response = validate_and_get_user(current_app)
+    elif auth_method == "shibboleth":
+        user, error_response = validate_sso_user(request)
+    else:
+        return False
+
+    return user is not None and error_response is None
+
+def register_before_request_auth(app):
+    """Attach the before_request auth logic to the Flask app."""
+
+    whitelist_routes = [
+        '/login_dev',
+        '/login_sso',
+        '/login',
+        '/callback',
+        '/favicon.ico',
+        '/flask/js/',
+        '/static',
+        '/flask/assets',
+        '/flask/img'
+    ]
+
+    @app.before_request
+    def enforce_authentication():
+        ENABLE_AUTH = app.config.get("ENABLE_AUTH", False)
+
+        if not ENABLE_AUTH:
+            return None
+
+        requested_path = request.path
+        if any(requested_path.startswith(route) for route in whitelist_routes):
+            return None
+
+        if not is_authenticated():
+            redirect_uri = app.config.get("LOGIN_REDIRECT_URL", "/login_dev")
+            print(f"Unauthorized access to {requested_path}. Redirecting.")
+            return redirect(redirect_uri)
+
+        return None
+
+def sync_auth0_users_to_db(app):
+    from mdvtools.dbutils.dbservice import UserService, UserProjectService
+    from mdvtools.dbutils.dbmodels import db, Project
+    from auth0.management import Auth0
+    from auth0.authentication import GetToken
+
+    """
+    Syncs users from Auth0 to the application's database using UserService and UserProjectService.
+    """
+    try:
+        # Load Auth0 config from app
+        auth0_domain = app.config['AUTH0_DOMAIN']
+        client_id = app.config['AUTH0_CLIENT_ID']
+        client_secret = app.config['AUTH0_CLIENT_SECRET']
+        auth0_db_connection = app.config['AUTH0_DB_CONNECTION']
+        audience = f"https://{auth0_domain}/api/v2/"
+
+        # Get Auth0 Management API token
+        get_token = GetToken(domain=auth0_domain, client_id=client_id, client_secret=client_secret)
+        mgmt_api_token = get_token.client_credentials(audience=audience)["access_token"]
+        auth0 = Auth0(auth0_domain, mgmt_api_token)
+
+        # Fetch users from Auth0 connection
+        users = auth0.users.list(q=f'identities.connection:"{auth0_db_connection}"', per_page=100)
+
+        for user in users['users']:
+            email = user.get('email', '')
+            auth0_id = user['user_id']
+
+            print(f"Syncing user: {email} - {auth0_id}")
+
+            # Use UserService to add or update user
+            db_user = UserService.add_or_update_user(
+                email=email,
+                auth0_id=auth0_id
+            )
+
+            # Fetch user's roles to determine admin status
+            roles = auth0.users.list_roles(auth0_id)
+            is_admin = any(role['name'] == 'admin' for role in roles['roles'])
+
+            # Update admin status
+            db_user.is_admin = is_admin
+            db.session.commit()
+
+            if is_admin:
+                # Assign all projects to this user as owner via UserProjectService
+                for project in Project.query.all():
+                    UserProjectService.add_or_update_user_project(
+                        user_id=db_user.id,
+                        project_id=project.id,
+                        is_owner=True
+                    )
+
+        print(f"Synced {len(users['users'])} users from Auth0.")
+
+    except Exception as e:
+        print(f"sync_auth0_users_to_db: An unexpected error occurred: {e}")
+        raise
+
+def cache_user_projects():
+    from mdvtools.dbutils.dbmodels import User, UserProject
+    from mdvtools.dbutils.dbservice import ProjectService
+    """
+    Caches user details and their associated project permissions in memory.
+    """
+    try:
+        print("Caching user details and project permissions...")
+
+        users = User.query.all()
+        all_users_cache.clear()
+        user_cache.clear()
+        user_project_cache.clear()
+        active_projects_cache.clear()
+
+        for user in users:
+            user_data = {
+                "id": user.id,
+                "auth0_id": user.auth0_id,
+                "email": user.email,
+                "is_admin": user.is_admin
+            }
+            user_cache[user.auth0_id] = user_data
+            all_users_cache.append(user_data)
+
+            user_projects = UserProject.query.filter_by(user_id=user.id).all()
+            project_permissions = {
+                up.project_id: {
+                    "can_read": up.can_read,
+                    "can_write": up.can_write,
+                    "is_owner": up.is_owner
+                }
+                for up in user_projects
+            }
+            user_project_cache[user.id] = project_permissions
+
+        active_projects = ProjectService.get_active_projects()
+        active_projects_cache[:] = active_projects
+
+        print(f"Cached {len(users)} users and their project permissions in memory.")
+        return True
+
+    except Exception as e:
+        print(f"Error caching user projects: {e}")
+        return False
+    
+def update_cache(user_id=None, project_id=None, user_data=None, project_data=None, permissions=None):
+    """
+    Updates the in-memory caches (user_cache, user_project_cache, all_users_cache, active_projects_cache)
+    for the provided user and project details, including any changes in permissions.
+
+    :param user_id: The ID of the user whose cache needs to be updated.
+    :param project_id: The ID of the project to be added or updated in the cache.
+    :param user_data: A dictionary containing the user's details (if updating the user cache).
+    :param project_data: A dictionary containing the project's details (if updating the project cache).
+    :param permissions: A dictionary containing the user's permissions for a project (optional).
+    """
+    try:
+        # Step 1: Update user_cache (user details)
+        if user_data and user_id:
+            user_cache[user_id] = user_data  # Update or add the user in the cache
+            
+            # Ensure the user is in all_users_cache if they are not already
+            if user_data not in all_users_cache:
+                all_users_cache.append(user_data)
+        
+        # Step 2: Update user_project_cache (user-project relationship)
+        if user_id and project_id and permissions is not None:
+            # Create or update user project permissions in cache
+            if user_id not in user_project_cache:
+                user_project_cache[user_id] = {}
+
+            user_project_cache[user_id][project_id] = permissions  # Update or add project permissions
+            
+        # Step 3: Update active_projects_cache (active project details)
+        if project_data and project_id:
+            # Search for the project in the cache
+            existing_project = next((p for p in active_projects_cache if p["id"] == project_id), None)
+
+            if existing_project:
+                # Project exists in cache, update only the changed fields
+                existing_project["name"] = project_data.get("name", existing_project["name"])
+                existing_project["lastModified"] = project_data.get("lastModified", existing_project["lastModified"])
+                existing_project["thumbnail"] = project_data.get("thumbnail", existing_project["thumbnail"])
+                print(f"Updated project {project_id} in active projects cache.")
+            else:
+                # Project does not exist in cache, append a new entry
+                project_entry = {
+                    "id": project_data["id"],
+                    "name": project_data["name"],
+                    "lastModified": project_data["lastModified"],
+                    "thumbnail": project_data["thumbnail"]
+                }
+                active_projects_cache.append(project_entry)
+                print(f"Added new project {project_id} to active projects cache.")
+
+        print("Cache successfully updated.")
+    
+    except Exception as e:
+        print(f"Error updating cache: {e}")
 
 def validate_and_get_user():
     from mdvtools.auth.auth0_provider import Auth0Provider
     from mdvtools.dbutils.mdv_server_app import oauth
-    from mdvtools.dbutils.mdv_server_app import user_cache
     from mdvtools.dbutils.dbmodels import User
     """
     Validates the Auth0 token from the session and retrieves the user from cache.
@@ -75,7 +291,6 @@ def validate_and_get_user():
     
 
 def validate_sso_user(request):
-    from mdvtools.dbutils.mdv_server_app import user_cache
     """
     Validates the SSO-authenticated user by checking required headers.
     Returns the user info from the cache or DB, or an error response.
